Corrected bug in box/case/pallet analysis...
* Changed version to 1.0.37.0,
* Corrected problem has pointed out by user Frédéric COVASSO : in file
"PalletSolutionDB\index.xml" , case and pallet weight might be saved
using comma (,)as decimal separator depending on windows regional
settings. When loaded, the comma is ignored and the weight values are
wrong. --> as a consequence, users using the PalletSolutionDB need to
edit their index.xml file and replace comma with dot in weight values.